+++ subprojects/github-publish/src/main/kotlin/github/publish/GitHubPublicationFactory.kt
@@ -1,7 +1,5 @@
 package com.nisecoder.gradle.plugin.github.publish
 
-import com.nisecoder.gradle.plugin.github.publish.DefaultGitHubPublication
-import com.nisecoder.gradle.plugin.github.publish.GitHubPublication
 import org.gradle.api.NamedDomainObjectFactory
 import org.gradle.api.model.ObjectFactory
 import org.gradle.kotlin.dsl.newInstance
@@ -16,6 +14,6 @@ class GitHubPublicationFactory(
      * @return The object.
      */
     override fun create(name: String): GitHubPublication {
-        return objectFactory.newInstance(DefaultGitHubPublication::class)
+        return objectFactory.newInstance(GitHubPublication::class)
     }
 }
